Pass scheduled time to jobs as properties.
Notes on this change:

- We can now grab an ExecutableFlow's override properties. This is important because:
- scheduleNow() (renamed to restartFlow for clarity) has to grab a job's scheduled time
  when it restarts a job. Previously, restarted flows had a scheduled execution time of
  whenever they were restarted. If jobs rely on that scheduled execution time, then things
  break when they're restarted.

The scheduled execution time should always be consistent with when the job was intended to
run. If we ever have a need to keep track of the time that the scheduler *eventually* scheduled
the job (i.e. moved it to the present because it was scheduled in the past), that should be
an entirely separate notion in the Scheduler.